Rename the product-page DOM readers to variable-product-page to match their scope
