+++ client/express-checkout/utils/index.js
@@ -14,7 +14,7 @@ import {
 
 export * from './normalize';
 export * from './bookings';
-export * from './product-page';
+export * from './variable-product-page';
 
 /**
  * Get error messages from WooCommerce notice.
